Resolved inode read / write synch problems.
Sometimes syn-rw test in filesys/extend fails, and by some inspection
read / write synch is cause of the problem.
By some edge case of process interleaving, Read operation in child-syn
is former than write operation in parent process.
So i added inode lock in inode and acquiring and relasing that in
read / write function and hence resolved it.

But i did not implement reader writer lock yet. So it's performance
may not good.